+++ ftkGUI/LabelImageViewQT.h
@@ -122,7 +122,7 @@ public:
 	void DoubleClicksOn(void){ enableDoubleClicks = true;};
 	//5D Image;
 	int GetCurrentTimeVal(void);
-	void SetCurrentTimeVal(int time);
+	void SetCurrentTimeVal(double time);
 
 		
 public slots:
